requirements.txt
okx-api-v5==1.0.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
ta-lib==0.4.28
scikit-learn==1.3.0
matplotlib==3.7.2 